[PHP8.5] Use canonical type casts

+++ CRM/Utils/Chart.php
@@ -99,7 +99,7 @@ class CRM_Utils_Chart {
     foreach ($params['multiValues'] as $i => $dataSet) {
       $output['values'][$i] = [];
       foreach ($dataSet as $k => $v) {
-        $output['values'][$i][] = ['label' => $k, 'value' => (double) $v];
+        $output['values'][$i][] = ['label' => $k, 'value' => (float) $v];
       }
     }
     if (!$output['values']) {
